Restructure `Header` component namespace
After adding responsive menu the .children file became lengthy. I
decided to restructure the namespace dividing components that form the
`Header` component into separate files.
