+++ src/components/Header/index.ts
@@ -1,2 +1,3 @@
 export { default as Header } from "./Header"
 export type { HeaderProps } from "./Header"
+export type { NavigationMenuItemType } from "./NavigationMenu"
